+++ borsh-rs/borsh-schema-derive-internal/src/enum_schema.rs
@@ -104,10 +104,10 @@ pub fn process_enum(input: &ItemEnum, cratename: Ident) -> syn::Result<TokenStre
     }
 
     let type_definitions = quote! {
-        fn add_definitions_recursively(definitions: &mut #cratename::lib::HashMap<#cratename::schema::Declaration, #cratename::schema::Definition>) {
+        fn add_definitions_recursively(definitions: &mut #cratename::maybestd::collections::HashMap<#cratename::schema::Declaration, #cratename::schema::Definition>) {
             #anonymous_defs
             #add_recursive_defs
-            let variants = #cratename::lib::vec![#(#variants_defs),*];
+            let variants = #cratename::maybestd::vec![#(#variants_defs),*];
             let definition = #cratename::schema::Definition::Enum{variants};
             Self::add_definition(Self::declaration(), definition, definitions);
         }
@@ -153,7 +153,7 @@ mod tests {
                     "A".to_string()
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
@@ -164,7 +164,7 @@ mod tests {
                     struct AEggs;
                     <ABacon>::add_definitions_recursively(definitions);
                     <AEggs>::add_definitions_recursively(definitions);
-                    let variants = borsh::lib::vec![
+                    let variants = borsh::maybestd::vec![
                         ("Bacon".to_string(), <ABacon>::declaration()),
                         ("Eggs".to_string(), <AEggs>::declaration())
                     ];
@@ -191,7 +191,7 @@ mod tests {
                     "A".to_string()
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
@@ -199,7 +199,7 @@ mod tests {
                     #[derive(borsh :: BorshSchema)]
                     struct ABacon;
                     <ABacon>::add_definitions_recursively(definitions);
-                    let variants = borsh::lib::vec![("Bacon".to_string(), <ABacon>::declaration())];
+                    let variants = borsh::maybestd::vec![("Bacon".to_string(), <ABacon>::declaration())];
                     let definition = borsh::schema::Definition::Enum { variants };
                     Self::add_definition(Self::declaration(), definition, definitions);
                 }
@@ -226,7 +226,7 @@ mod tests {
                     "A".to_string()
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
@@ -246,7 +246,7 @@ mod tests {
                     <AEggs>::add_definitions_recursively(definitions);
                     <ASalad>::add_definitions_recursively(definitions);
                     <ASausage>::add_definitions_recursively(definitions);
-                    let variants = borsh::lib::vec![
+                    let variants = borsh::maybestd::vec![
                         ("Bacon".to_string(), <ABacon>::declaration()),
                         ("Eggs".to_string(), <AEggs>::declaration()),
                         ("Salad".to_string(), <ASalad>::declaration()),
@@ -279,11 +279,11 @@ mod tests {
                 W: borsh::BorshSchema
             {
                 fn declaration() -> borsh::schema::Declaration {
-                    let params = borsh::lib::vec![<C>::declaration(), <W>::declaration()];
+                    let params = borsh::maybestd::vec![<C>::declaration(), <W>::declaration()];
                     format!(r#"{}<{}>"#, "A", params.join(", "))
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
@@ -310,7 +310,7 @@ mod tests {
                     <AEggs<C, W> >::add_definitions_recursively(definitions);
                     <ASalad<C, W> >::add_definitions_recursively(definitions);
                     <ASausage<C, W> >::add_definitions_recursively(definitions);
-                    let variants = borsh::lib::vec![
+                    let variants = borsh::maybestd::vec![
                         ("Bacon".to_string(), <ABacon<C, W> >::declaration()),
                         ("Eggs".to_string(), <AEggs<C, W> >::declaration()),
                         ("Salad".to_string(), <ASalad<C, W> >::declaration()),
