+++ borsh-rs/borsh-schema-derive-internal/src/struct_schema.rs
@@ -36,7 +36,7 @@ pub fn process_struct(input: &ItemStruct, cratename: Ident) -> syn::Result<Token
             }
             if !fields_vec.is_empty() {
                 struct_fields = quote! {
-                    let fields = #cratename::schema::Fields::NamedFields(#cratename::lib::vec![#(#fields_vec),*]);
+                    let fields = #cratename::schema::Fields::NamedFields(#cratename::maybestd::vec![#(#fields_vec),*]);
                 };
             }
         }
@@ -58,7 +58,7 @@ pub fn process_struct(input: &ItemStruct, cratename: Ident) -> syn::Result<Token
             }
             if !fields_vec.is_empty() {
                 struct_fields = quote! {
-                    let fields = #cratename::schema::Fields::UnnamedFields(#cratename::lib::vec![#(#fields_vec),*]);
+                    let fields = #cratename::schema::Fields::UnnamedFields(#cratename::maybestd::vec![#(#fields_vec),*]);
                 };
             }
         }
@@ -72,7 +72,7 @@ pub fn process_struct(input: &ItemStruct, cratename: Ident) -> syn::Result<Token
     }
 
     let add_definitions_recursively = quote! {
-        fn add_definitions_recursively(definitions: &mut #cratename::lib::HashMap<#cratename::schema::Declaration, #cratename::schema::Definition>) {
+        fn add_definitions_recursively(definitions: &mut #cratename::maybestd::collections::HashMap<#cratename::schema::Declaration, #cratename::schema::Definition>) {
             #struct_fields
             let definition = #cratename::schema::Definition::Struct { fields };
             Self::add_definition(Self::declaration(), definition, definitions);
@@ -118,7 +118,7 @@ mod tests {
                 fn declaration() -> borsh::schema::Declaration {
                     "A".to_string()
                 }
-                fn add_definitions_recursively(definitions: &mut borsh::lib::HashMap<borsh::schema::Declaration, borsh::schema::Definition>) {
+                fn add_definitions_recursively(definitions: &mut borsh::maybestd::collections::HashMap<borsh::schema::Declaration, borsh::schema::Definition>) {
                     let fields = borsh::schema::Fields::Empty;
                     let definition = borsh::schema::Definition::Struct { fields };
                     Self::add_definition(Self::declaration(), definition, definitions);
@@ -143,16 +143,16 @@ mod tests {
                 T: borsh::BorshSchema
             {
                 fn declaration() -> borsh::schema::Declaration {
-                    let params = borsh::lib::vec![<T>::declaration()];
+                    let params = borsh::maybestd::vec![<T>::declaration()];
                     format!(r#"{}<{}>"#, "A", params.join(", "))
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
                 ) {
-                    let fields = borsh::schema::Fields::UnnamedFields(borsh::lib::vec![<T>::declaration()]);
+                    let fields = borsh::schema::Fields::UnnamedFields(borsh::maybestd::vec![<T>::declaration()]);
                     let definition = borsh::schema::Definition::Struct { fields };
                     Self::add_definition(Self::declaration(), definition, definitions);
                     <T>::add_definitions_recursively(definitions);
@@ -180,12 +180,12 @@ mod tests {
                     "A".to_string()
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
                 ) {
-                    let fields = borsh::schema::Fields::UnnamedFields(borsh::lib::vec![
+                    let fields = borsh::schema::Fields::UnnamedFields(borsh::maybestd::vec![
                         <u64>::declaration(),
                         <String>::declaration()
                     ]);
@@ -216,17 +216,17 @@ mod tests {
                 V: borsh::BorshSchema
             {
                 fn declaration() -> borsh::schema::Declaration {
-                    let params = borsh::lib::vec![<K>::declaration(), <V>::declaration()];
+                    let params = borsh::maybestd::vec![<K>::declaration(), <V>::declaration()];
                     format!(r#"{}<{}>"#, "A", params.join(", "))
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
                 ) {
                     let fields =
-                        borsh::schema::Fields::UnnamedFields(borsh::lib::vec![<K>::declaration(), <V>::declaration()]);
+                        borsh::schema::Fields::UnnamedFields(borsh::maybestd::vec![<K>::declaration(), <V>::declaration()]);
                     let definition = borsh::schema::Definition::Struct { fields };
                     Self::add_definition(Self::declaration(), definition, definitions);
                     <K>::add_definitions_recursively(definitions);
@@ -258,12 +258,12 @@ mod tests {
                     "A".to_string()
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
                 ) {
-                    let fields = borsh::schema::Fields::NamedFields(borsh::lib::vec![
+                    let fields = borsh::schema::Fields::NamedFields(borsh::maybestd::vec![
                         ("x".to_string(), <u64>::declaration()),
                         ("y".to_string(), <String>::declaration())
                     ]);
@@ -297,16 +297,16 @@ mod tests {
                 String: borsh::BorshSchema
             {
                 fn declaration() -> borsh::schema::Declaration {
-                    let params = borsh::lib::vec![<K>::declaration(), <V>::declaration()];
+                    let params = borsh::maybestd::vec![<K>::declaration(), <V>::declaration()];
                     format!(r#"{}<{}>"#, "A", params.join(", "))
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
                 ) {
-                    let fields = borsh::schema::Fields::NamedFields(borsh::lib::vec![
+                    let fields = borsh::schema::Fields::NamedFields(borsh::maybestd::vec![
                         ("x".to_string(), <HashMap<K, V> >::declaration()),
                         ("y".to_string(), <String>::declaration())
                     ]);
@@ -334,7 +334,7 @@ mod tests {
                     "A".to_string()
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
@@ -365,12 +365,12 @@ mod tests {
                     "A".to_string()
                 }
                 fn add_definitions_recursively(
-                    definitions: &mut borsh::lib::HashMap<
+                    definitions: &mut borsh::maybestd::collections::HashMap<
                         borsh::schema::Declaration,
                         borsh::schema::Definition
                     >
                 ) {
-                    let fields = borsh::schema::Fields::UnnamedFields(borsh::lib::vec![<String>::declaration()]);
+                    let fields = borsh::schema::Fields::UnnamedFields(borsh::maybestd::vec![<String>::declaration()]);
                     let definition = borsh::schema::Definition::Struct { fields };
                     Self::add_definition(Self::declaration(), definition, definitions);
                     <String>::add_definitions_recursively(definitions);
