+++ borsh-rs/borsh/src/schema_helpers.rs
@@ -1,6 +1,7 @@
-use crate::lib::*;
-
-use crate::error::{Error, ErrorKind, Result};
+use crate::maybestd::{
+    io::{Error, ErrorKind, Result},
+    vec::Vec
+};
 use crate::schema::BorshSchemaContainer;
 use crate::{BorshDeserialize, BorshSchema, BorshSerialize};
