+++ borsh-rs/borsh/tests/test_schema_enums.rs
@@ -1,12 +1,13 @@
 #![allow(dead_code)]  // Local structures do not have their fields used.
 use borsh::schema::*;
 use borsh::schema_helpers::{try_from_slice_with_schema, try_to_vec_with_schema};
+use borsh::maybestd::collections::HashMap;
 
 macro_rules! map(
-    () => { borsh::lib::HashMap::new() };
+    () => { HashMap::new() };
     { $($key:expr => $value:expr),+ } => {
         {
-            let mut m = borsh::lib::HashMap::new();
+            let mut m = HashMap::new();
             $(
                 m.insert($key.to_string(), $value);
             )+
