+++ borsh-rs/borsh/tests/test_schema_nested.rs
@@ -1,11 +1,12 @@
 #![allow(dead_code)]  // Local structures do not have their fields used.
 use borsh::schema::*;
+use borsh::maybestd::collections::HashMap;
 
 macro_rules! map(
-    () => { borsh::lib::HashMap::new() };
+    () => { HashMap::new() };
     { $($key:expr => $value:expr),+ } => {
         {
-            let mut m = borsh::lib::HashMap::new();
+            let mut m = HashMap::new();
             $(
                 m.insert($key.to_string(), $value);
             )+
@@ -23,7 +24,7 @@ pub fn duplicated_instantiations() {
     struct Cucumber;
     #[derive(borsh::BorshSchema)]
     struct Oil<K, V> {
-        seeds: borsh::lib::HashMap<K, V>,
+        seeds: HashMap<K, V>,
         liquid: Option<K>,
     };
     #[derive(borsh::BorshSchema)]
