+++ borsh-rs/borsh/tests/test_schema_structs.rs
@@ -1,10 +1,11 @@
 use borsh::schema::*;
+use borsh::maybestd::collections::HashMap;
 
 macro_rules! map(
-    () => { borsh::lib::HashMap::new() };
+    () => { HashMap::new() };
     { $($key:expr => $value:expr),+ } => {
         {
-            let mut m = borsh::lib::HashMap::new();
+            let mut m = HashMap::new();
             $(
                 m.insert($key.to_string(), $value);
             )+
@@ -105,7 +106,7 @@ pub fn tuple_struct_params() {
 pub fn simple_generics() {
     #[derive(borsh::BorshSchema)]
     struct A<K, V> {
-        _f1: borsh::lib::HashMap<K, V>,
+        _f1: HashMap<K, V>,
         _f2: String,
     };
     assert_eq!(
