+++ borsh-rs/borsh/tests/test_simple_structs.rs
@@ -1,5 +1,5 @@
 use borsh::{BorshDeserialize, BorshSerialize};
-use borsh::lib::{HashMap, HashSet};
+use borsh::maybestd::collections::{HashMap, HashSet};
 
 #[derive(BorshSerialize, BorshDeserialize, PartialEq, Debug)]
 #[borsh_init(init)]
